requirements.txt
functions-framework==3.*
google-cloud-vision
google-api-python-client
google-auth
firebase-admin